Added shasum for Raspberry Pi binary to attributes/default.rb
Installing from a binary was failing on a Raspberry Pi due to a null string error. I added the shasum to the attributes/default.rb file. However, I used the shasum that I calculated locally. This is wrong. I tried to go to http://nodejs.org/dist/v0.10.15/ and get the shasum posted there from the gpg or asc file, however none of those checksum files, include the necessary checksum for the "arm-pi" flavor of NodeJS.

In the absence of an official shasum, can someone find out how to get the tarball over https and confirm the shasum above before merging this pull request? I pulled the tarball over http and therefore my shasum should be treated as suspect until independently verified. In the meantime, I filed an issue over in the joyent/node repo:

https://github.com/joyent/node/issues/6338