Remove iojs installation and use nodesource.com on RHEL not epel
EPEL is ancient on RHEL 6 to the point that’s it’s useless. Using nodesource aligns us with Debian based systems and removes another cookbook dep while we’re at it (yum-epel)

iojs is a dead project and nodejs 4 includes all the same features. There’s no point in continuing to support it now

Signed-off-by: Tim Smith <tsmith@chef.io>