Convert nodejs_npm to a custom resource and rename npm_package
If we ever pull this into core chef it will be npm_package so we should
start with the logical name now. I've retained backwards compatibility
with the existing resource name, but people should probably update their
cookbooks eventually to use the new name. This is also a custom resource
now and I've removed the use of compat_resource since it's incredibly
problematic and instead this cookbook now requires Chef 12.14 or later
due to the use of yum_repository. 12.14 shipped 14 months ago at this
point so I don't think it's too much to ask that people have migrated to
that version.

Signed-off-by: Tim Smith <tsmith@chef.io>